download, remote copy, stop refactor

+++ manage_server.go
@@ -3,21 +3,26 @@ package main
 import (
 	"encoding/base64"
 	"fmt"
+	"io/ioutil"
 	"log"
+
+	"github.com/pkg/sftp"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/ssm"
+	"golang.org/x/crypto/ssh"
 )
 
 const startupScript = `#!/bin/bash
+				       cd /home/ec2-user/
 					   yum update -y
 					   yum install java-1.8.0-openjdk-headless.x86_64 -y
-					   wget https://launcher.mojang.com/v1/objects/3dc3d84a581f14691199cf6831b71ed1296a9fdf/server.jar -O minecraft.jar
-					   java -Xmx1024M -Xms512M -jar minecraft.jar nogui
+					   runuser -l ec2-user -c 'wget https://launcher.mojang.com/v1/objects/3dc3d84a581f14691199cf6831b71ed1296a9fdf/server.jar -O minecraft.jar'
+					   runuser -l ec2-user -c 'java -Xms512M -Xmx1024M -jar minecraft.jar nogui'
 					   sed -i -e s/eula=false/eula=true/g eula.txt
-					   screen -dmS minecraft java -Xmx1024M -Xms512M -jar minecraft.jar nogui`
+					   runuser -l ec2-user -c 'screen -dmS minecraft java -Xmx1024M -Xms512M -jar minecraft.jar nogui'`
 
 func start(session *session.Session, tagKey, tagValue string) (ipAddress string) {
 	svc := ec2.New(session)
@@ -36,7 +41,8 @@ func start(session *session.Session, tagKey, tagValue string) (ipAddress string)
 		InstanceType:   aws.String("t2.micro"),
 		MinCount:       aws.Int64(1),
 		MaxCount:       aws.Int64(1),
-		SecurityGroups: []*string{aws.String(tagValue)},
+		SecurityGroups: []*string{aws.String(tagValue), aws.String("allow-ssh")},
+		KeyName:        aws.String("myFirstKey"),
 		UserData:       aws.String(base64.StdEncoding.EncodeToString([]byte(startupScript))),
 	})
 	if err != nil {
@@ -45,6 +51,7 @@ func start(session *session.Session, tagKey, tagValue string) (ipAddress string)
 	}
 	mcInstanceId := *runResult.Instances[0].InstanceId
 
+	fmt.Println("Waiting for instance to boot...")
 	err = svc.WaitUntilInstanceRunning(&ec2.DescribeInstancesInput{
 		InstanceIds: []*string{aws.String(mcInstanceId)},
 	})
@@ -59,8 +66,7 @@ func start(session *session.Session, tagKey, tagValue string) (ipAddress string)
 		panic(err)
 	}
 
-	// does not exist yet, need to wait
-	fmt.Println("Created instance:", *minecraftInstance.Reservations[0].Instances[0].PublicIpAddress)
+	fmt.Printf("Created instance: %s\n", *minecraftInstance.Reservations[0].Instances[0].InstanceId)
 
 	_, errtag := svc.CreateTags(&ec2.CreateTagsInput{
 		Resources: []*string{runResult.Instances[0].InstanceId},
@@ -106,16 +112,19 @@ func fetchRunningInstancesByTag(session *session.Session, tagKey, tagValue strin
 	return instances
 }
 
-func stop(session *session.Session, tagValue string) {
+func stop(session *session.Session, tagKey, tagValue string) {
 	svc := ec2.New(session)
 	instancesToDelete := fetchRunningInstancesByTag(session, tagKey, tagValue)
+	if len(instancesToDelete) == 0 {
+		fmt.Println("No instances to delete")
+		return
+	}
 
 	var instanceIds []string
 	for _, instance := range instancesToDelete {
 		instanceIds = append(instanceIds, *instance.InstanceId)
 	}
 
-	fmt.Println(instancesToDelete)
 	terminateOutput, err := svc.TerminateInstances(&ec2.TerminateInstancesInput{
 		InstanceIds: aws.StringSlice(instanceIds),
 	})
@@ -123,20 +132,80 @@ func stop(session *session.Session, tagValue string) {
 		panic(err)
 	}
 	for _, instanceStateChange := range terminateOutput.TerminatingInstances {
-		fmt.Printf("Instance %s %s\n", *instanceStateChange.InstanceId, *instanceStateChange.CurrentState)
+		fmt.Printf("Instance %s %s\n", *instanceStateChange.InstanceId, *instanceStateChange.CurrentState.Name)
 	}
 }
 
-func test(session *session.Session) {
+func downloadWorld() error {
+	/*
+		config := &ssh.ClientConfig{
+			User: "ec2-user",
+			Auth: []ssh.AuthMethod{
+				publicKey("myFirstKey"),
+			},
+			HostKeyCallback: ssh.InsecureIgnoreHostKey(),
+		}
+	*/
+	conn, err := ssh.Dial("tcp", "52.209.252.105:22", sshConfig)
+	if err != nil {
+		panic(err)
+	}
+
+	sftpClient, err := sftp.NewClient(conn)
+	if err != nil {
+		panic(err)
+	}
+	defer sftpClient.Close()
+
+	err = copyRemoteDirToLocal("/home/ec2-user/world", "world", sftpClient)
+	if err != nil {
+		return fmt.Errorf("Could not download world %v", err)
+	}
+	/*err = copyLocalDirToRemote("cake", "/home/ec2-user", sftpClient)
+	if err != nil {
+		return fmt.Errorf("Could not upload dir %v", err)
+	}*/
+	return nil
+}
+
+func publicKey(path string) ssh.AuthMethod {
+	key, err := ioutil.ReadFile("/home/jonas/.ssh/myFirstKey.pem")
+	if err != nil {
+		panic(err)
+	}
+	signer, err := ssh.ParsePrivateKey(key)
+	if err != nil {
+		panic(err)
+	}
+	return ssh.PublicKeys(signer)
+}
+
+func stopMcServer() {
+	conn, err := ssh.Dial("tcp", "52.214.47.40:22", sshConfig)
+	if err != nil {
+		panic(err)
+	}
+
+	sess, err := conn.NewSession()
+	if err != nil {
+		panic(err)
+	}
+
+	err = sess.Run(`screen -S minecraft -p 0 -X stuff "stop^M"`)
+	if err != nil {
+		panic(err)
+	}
+}
+
+func ssmStopMcServer(session *session.Session, instanceId string) {
 	svc := ssm.New(session)
 
 	sendcommandOutput, err := svc.SendCommand(&ssm.SendCommandInput{
 		DocumentName: aws.String("AWS-RunShellScript"),
 		Parameters: map[string][]*string{"commands": []*string{
-			aws.String("echo more-hello"),
-			aws.String("echo hello-again"),
+			aws.String(`screen -S minecraft -p 0 -X stuff "stop^M"`),
 		}},
-		InstanceIds: []*string{aws.String("i-0041d402a819f00b3")},
+		InstanceIds: []*string{aws.String(instanceId)},
 	})
 	if err != nil {
 		panic(err)
